preparing shift from dictionary package to namespace package

+++ cnf/src/main/java/cnuphys/cnf/export/RecParticleExporter.java
@@ -1,17 +1,16 @@
 package cnuphys.cnf.export;
 
 import java.io.FileNotFoundException;
-
-import org.jlab.clas.physics.Particle;
-import org.jlab.detector.base.DetectorType;
-import org.jlab.io.base.DataBank;
-import org.jlab.io.base.DataEvent;
-
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.UnsupportedEncodingException;
+
+import org.jlab.clas.physics.Particle;
+import org.jlab.detector.base.DetectorType;
+import org.jlab.io.base.DataBank;
+import org.jlab.io.base.DataEvent;
 
 
 //import cnuphys.cnf.event.dictionary.Bank;
@@ -19,11 +18,11 @@ import java.io.UnsupportedEncodingException;
 //import cnuphys.cnf.event.dictionary.Dictionary;
 
 public class RecParticleExporter extends AExporter {
-	
+
 	//-----IMPORTANT VALUES-----//
     private final static double targetMass = 0.938272;
     private final static double beamEnergy = 6.535;
- 
+
 
 	//the data output stream
 	//private DataOutputStream _dos;
@@ -68,39 +67,39 @@ public class RecParticleExporter extends AExporter {
 
 	@Override
 	public void nextEvent(DataEvent event) {
-		
-		
+
+
 		DataBank recPart = null;
 		DataBank recTrack = null;
 		DataBank recTraj = null;
 		DataBank recScin = null;
 		DataBank recCal = null;
 		DataBank recCher = null;
-		
+
 		if(event.hasBank("REC::Particle")) 		recPart = event.getBank("REC::Particle");
 		if(event.hasBank("REC::Track"))			recTrack = event.getBank("REC::Track");
 		if(event.hasBank("REC::Traj"))			recTraj = event.getBank("REC::Traj");
 		if(event.hasBank("REC::Scintillator"))	recScin = event.getBank("REC::Scintillator");
 		if(event.hasBank("REC::Calorimeter"))	recCal = event.getBank("REC::Calorimeter");
 		if(event.hasBank("REC::Cherenkov"))		recCher = event.getBank("REC::Cherenkov");
-		
-		
+
+
 		if (recPart == null || recTrack == null || recTraj == null || recScin == null || recCal == null || recCher == null) return;
-		
-		
-		
+
+
+
 		Particle beam_Part = new Particle(11, 0, 0, beamEnergy, 0, 0, 0);
         Particle target_Part = Particle.createWithMassCharge(targetMass, +1, 0,0,0, 0,0,0);
         Particle electron = null;
         Particle proton = null;
-	
+
         for(int ii = 0; ii < recPart.rows(); ii++) {
         	int pid = recPart.getInt("pid", ii);
-        	
+
         	if (pid == 0) {
         		continue;
         	}
-        	
+
 			float chi2pid = recPart.getFloat("chi2pid", ii);
             int status = Math.abs(recPart.getShort("status", ii));
             //if status_abs == 2 then particle is in FD but if status_abs == 4 then particle is in CD
@@ -110,7 +109,7 @@ public class RecParticleExporter extends AExporter {
     		float px = recPart.getFloat("px", ii);
 			float py = recPart.getFloat("py", ii);
 			float pz = recPart.getFloat("pz", ii);
-			
+
 			Particle part = new Particle(pid,
 					recPart.getFloat("px",ii),
 					recPart.getFloat("py",ii),
@@ -118,13 +117,13 @@ public class RecParticleExporter extends AExporter {
 					recPart.getFloat("vx",ii),
                     recPart.getFloat("vy",ii),
                     recPart.getFloat("vz",ii));
-			
+
 			double p = part.p();
-			
+
 			//electron is trigger particle (ii==0) and has correct pid
 			if(pid == 11 && ii == 0) {
 				electron = part;
-				
+
 				//check REC::Traj in the Drift Chambers
 				for(int it = 0; it < recTraj.rows(); it++) {
 					int pindex_traj = recTraj.getShort("pindex", it);
@@ -134,31 +133,31 @@ public class RecParticleExporter extends AExporter {
             		if (pindex_traj == ii) {
             			// check detector is DC
             			if (detector_traj == DetectorType.DC.getDetectorId()) {
-            				int region = (int) (layer_traj-1)/12 + 1;
+            				int region = (layer_traj-1)/12 + 1;
             				//System.out.printf("Region in DC: %d\n", region);
             			}
             		}
 				} //end of recTrack rows
 			} //end of check for electron
-			
+
 			//proton has correct pid
 			if(pid == 2212) {
 				proton = part;
 			} //end of check for proton
         }
-		
-		
+
+
 //		if (_columns == null) {
 //			return;
 //		}
-//		
+//
 //		//have data?
 //		if (!_recParticleBank.hasData(event)) {
 //			System.out.println("NO DATA");
 //			return;
 //		}
 //	//	System.out.println("HAS DATA");
-//		
+//
 		count++;
 		if ((count % 1000) == 0) {
 			System.err.println("Export count: " + count);
